small refactoring №2 of unit tests in dao, service and web module
