+++ service/src/test/java/com/fitgoal/service/RegistrationServiceImplTest.java
@@ -14,8 +14,8 @@ import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.Optional;
 
-import static com.fitgoal.service.util.TestHelper.createUserDto;
-import static com.fitgoal.service.util.TestHelper.createUserRegistrationData;
+import static com.fitgoal.service.util.ServiceTestHelper.createUserDto;
+import static com.fitgoal.service.util.ServiceTestHelper.createUserRegistrationData;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.mockito.Mockito.times;
