+++ service/src/test/java/com/fitgoal/service/util/ServiceTestHelper.java
@@ -1,10 +1,10 @@
-package com.fitgoal.web.resources.util;
+package com.fitgoal.service.util;
 
-import com.fitgoal.api.domain.User;
 import com.fitgoal.api.domain.UserLoginData;
 import com.fitgoal.api.domain.UserRegistrationData;
+import com.fitgoal.dao.domain.UserDto;
 
-public class TestHelper {
+public class ServiceTestHelper {
 
     public static UserLoginData createUserLoginData() {
         return UserLoginData.builder()
@@ -13,19 +13,20 @@ public class TestHelper {
                 .build();
     }
 
+    public static UserDto createUserDto() {
+        return UserDto.builder()
+                .id(1L)
+                .email("test@gmail.com")
+                .password("testPass")
+                .link("testLink")
+                .active(false)
+                .build();
+    }
+
     public static UserRegistrationData createUserRegistrationData() {
         return UserRegistrationData.builder()
                 .email("test@gmail.com")
                 .password("testPass")
-                .build();
-    }
-
-    public static User createUser() {
-        return User.builder()
-                .id(1L)
-                .email("test@gmail.com")
-                .link("testLink")
-                .active(false)
                 .build();
     }
 }
