+++ web/src/test/java/com/fitgoal/web/util/WebTestHelper.java
@@ -1,25 +1,15 @@
-package com.fitgoal.service.util;
+package com.fitgoal.web.util;
 
+import com.fitgoal.api.domain.User;
 import com.fitgoal.api.domain.UserLoginData;
 import com.fitgoal.api.domain.UserRegistrationData;
-import com.fitgoal.dao.domain.UserDto;
 
-public class TestHelper {
+public class WebTestHelper {
 
     public static UserLoginData createUserLoginData() {
         return UserLoginData.builder()
                 .email("test@gmail.com")
                 .password("testPass")
-                .build();
-    }
-
-    public static UserDto createUserDto() {
-        return UserDto.builder()
-                .id(1L)
-                .email("test@gmail.com")
-                .password("testPass")
-                .link("testLink")
-                .active(false)
                 .build();
     }
 
@@ -30,4 +20,12 @@ public class TestHelper {
                 .build();
     }
 
+    public static User createUser() {
+        return User.builder()
+                .id(1L)
+                .email("test@gmail.com")
+                .link("testLink")
+                .active(false)
+                .build();
+    }
 }
